Rgba<u8> -> Rgba<f64>

+++ src/intersection.rs
@@ -3,7 +3,7 @@ use na::Vector3;
 
 pub struct Intersection {
     pub pos: Vector3<f64>,
-    pub color: Rgba<u8>,
+    pub color: Rgba<f64>,
     pub opacity: f64,
     pub reflection: f64,
 }
